feat(60-01): relocate verified-resolve chokepoint to api-client with D-04 strict cutover
- Add cipherbox-core + ciborium deps to crates/api-client/Cargo.toml (enables decode_ipns_cbor_data)
- Move VerifyError (no Legacy), VerifiedResolve, bind_verified, resolve_ipns_verified to api-client/src/ipns.rs
- Remove VerifyError::Legacy variant (D-04) — all-absent sig fields now Err(Invalid)
- Remove first-publish skew disjunct (resp_seq==1 && embedded_seq==0) for strict equality (D-04)
- Remove Ok(None) all-absent branch from verify_ipns_resolve_signature (falls through to Some(false))
- Reduce crates/fuse/src/verify.rs to thin re-export of api-client symbols (D-08)
- Fold all 9 fuse Legacy caller arms into Invalid fail-closed handling (D-04 compile-time enforcement)
- All 15 api-client unit tests pass; cargo check -p cipherbox-fuse clean

Entire-Checkpoint: df13e20983da
