+++ crates/api-client/Cargo.toml
@@ -14,3 +14,5 @@ log = { workspace = true }
 base64 = { workspace = true }
 urlencoding = { workspace = true }
 cipherbox-crypto = { workspace = true }
+cipherbox-core = { workspace = true }
+ciborium = { workspace = true }
