+++ crates/fuse/src/metadata.rs
@@ -323,16 +323,7 @@ pub fn spawn_metadata_publish(
                     // D-01/D-02: route through verified chokepoint; fail only this merge operation.
                     let remote_cid = match crate::verify::resolve_ipns_verified(&api, &ipns_name).await {
                         Ok(v) => v.cid,
-                        Err(crate::verify::VerifyError::Legacy { cid, .. }) => {
-                            // D-04: legacy record — use the carried cid (no second resolve_ipns).
-                            // T-59-04: eliminates the TOCTOU race window.
-                            log::warn!(
-                                "remote_merge: IPNS {} resolved without signature fields \
-                                 — using DB CID for merge (D-04)",
-                                ipns_name
-                            );
-                            cid
-                        }
+                        // D-04: Legacy variant removed — all-absent sig fields fail closed.
                         Err(crate::verify::VerifyError::Invalid(msg)) => {
                             return Err(format!("IPNS {} verify failed on merge: {}", ipns_name, msg));
                         }
@@ -474,24 +465,7 @@ pub fn spawn_bin_entry_publish(
                             }
                         }
                     }
-                    Err(crate::verify::VerifyError::Legacy { cid, .. }) => {
-                        // D-04: legacy record — use the carried cid (no second resolve_ipns).
-                        // T-59-04: eliminates the TOCTOU race window.
-                        log::warn!(
-                            "spawn_bin_entry_publish: IPNS {} resolved without signature fields \
-                             — using DB CID (D-04)",
-                            bin_ipns_name
-                        );
-                        match cipherbox_api_client::ipfs::fetch_content(&api, &cid).await {
-                            Ok(bytes) => {
-                                match cipherbox_core::decrypt_bin_metadata(&bytes, &user_private_key) {
-                                    Ok(meta) => (meta, Some(cid)),
-                                    Err(e) => return Err(format!("Bin decrypt failed: {}", e)),
-                                }
-                            }
-                            Err(e) => return Err(format!("Bin fetch failed: {}", e)),
-                        }
-                    }
+                    // D-04: Legacy variant removed — all-absent sig fields fail closed.
                     Err(crate::verify::VerifyError::Invalid(msg)) => {
                         // D-02: fail only this operation.
                         log::warn!("spawn_bin_entry_publish: IPNS {} verify failed: {}", bin_ipns_name, msg);
@@ -642,16 +616,7 @@ async fn resolve_and_fetch_file_meta(
     // D-01: route through the verified chokepoint.
     let cid = match crate::verify::resolve_ipns_verified(api, file_meta_ipns_name).await {
         Ok(v) => v.cid,
-        Err(crate::verify::VerifyError::Legacy { cid, .. }) => {
-            // D-04: legacy record — use the carried cid (no second resolve_ipns).
-            // T-59-04: eliminates the TOCTOU race window.
-            log::warn!(
-                "resolve_and_fetch_file_meta: IPNS {} resolved without signature fields \
-                 — using DB CID (D-04)",
-                file_meta_ipns_name
-            );
-            cid
-        }
+        // D-04: Legacy variant removed — all-absent sig fields fail closed.
         Err(crate::verify::VerifyError::Invalid(msg)) => {
             // D-02: fail only this operation.
             return Err(format!(
