+++ crates/fuse/src/publish.rs
@@ -102,27 +102,8 @@ impl PublishCoordinator {
                 self.update_cache(ipns_name, seq);
                 Ok(seq)
             }
-            Err(crate::verify::VerifyError::Legacy { sequence_number, .. }) => {
-                // D-04: legacy record — use the carried sequence_number (no second resolve_ipns).
-                // T-59-04: carrying the already-classified response eliminates the TOCTOU race
-                // window where a concurrent publish could change the record in the ~1ms gap.
-                log::warn!(
-                    "resolve_sequence: IPNS {} resolved without signature fields \
-                     — using DB sequence (D-04)",
-                    ipns_name
-                );
-                let resolved = sequence_number.parse::<u64>().unwrap_or_else(|e| {
-                    log::warn!(
-                        "Failed to parse carried IPNS sequence '{}' for {}: {}",
-                        sequence_number, ipns_name, e
-                    );
-                    0
-                });
-                let cached = self.get_cached(ipns_name).unwrap_or(0);
-                let seq = std::cmp::max(resolved, cached);
-                self.update_cache(ipns_name, seq);
-                Ok(seq)
-            }
+            // D-04: Legacy variant removed — all-absent sig fields fail closed (strict cutover).
+            // Callers fall through to the Invalid arm.
             Err(crate::verify::VerifyError::Invalid(msg)) => {
                 // D-02: verify failure on soft resolve — fall back to cache (never wedge).
                 log::warn!(
@@ -170,23 +151,7 @@ impl PublishCoordinator {
                 self.update_cache(ipns_name, seq);
                 Ok(seq)
             }
-            Err(crate::verify::VerifyError::Legacy { sequence_number, .. }) => {
-                // D-04: legacy record — use the carried sequence_number (no second resolve_ipns).
-                // T-59-04: carrying the already-classified response eliminates the TOCTOU race
-                // window (strict contract: only fails on resolve failure, not legacy).
-                log::warn!(
-                    "resolve_sequence_strict: IPNS {} resolved without signature fields \
-                     — proceeding with DB sequence (D-04)",
-                    ipns_name
-                );
-                let resolved = sequence_number.parse::<u64>().map_err(|e| {
-                    format!("Invalid IPNS sequence '{}' for {}: {}", sequence_number, ipns_name, e)
-                })?;
-                let cached = self.get_cached(ipns_name).unwrap_or(0);
-                let seq = std::cmp::max(resolved, cached);
-                self.update_cache(ipns_name, seq);
-                Ok(seq)
-            }
+            // D-04: Legacy variant removed — all-absent sig fields fail closed (strict cutover).
             Err(crate::verify::VerifyError::Invalid(msg)) => {
                 // Strict: verify failure → Err.
                 Err(format!("IPNS {} verify failed: {}", ipns_name, msg))
